refactor: separate spacing property component
