+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Properties/blocks/Spacer/Spacing/index.ts
@@ -1,0 +1,1 @@
+export { Spacing } from './Spacing'
